# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface League {
  id: string;
  name: string;
  season: string;
  managerCount: number;
  currentGameweek: number;
  updatedAt: string;
  /** @nullable */
  fplLeagueId?: string | null;
}

export interface LeagueInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  season: string;
}

export interface FplConnectInput {
  /** @minLength 1 */
  fplLeagueId: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  season: string;
}

export interface Standing {
  rank: number;
  manager: string;
  teamName: string;
  gameweekPoints: number;
  totalPoints: number;
  previousRank: number;
  movement: number;
  initials?: string;
}

export interface HeadToHeadStanding {
  rank: number;
  manager: string;
  wins: number;
  losses: number;
  draws: number;
  pointsFor: number;
  pointsAgainst: number;
  pointDifference: number;
  leaguePoints: number;
  streak: string;
  initials?: string;
}

export interface Matchup {
  id: string;
  home: string;
  away: string;
  homeScore: number;
  awayScore: number;
  status: string;
  featured?: boolean;
}

export interface Activity {
  id: string;
  title: string;
  detail: string;
  time: string;
  kind: string;
}

export interface PowerRanking {
  rank: number;
  manager: string;
  movement: number;
  score: number;
}

export interface LeagueDashboard {
  league: League;
  topScorer: Standing;
  activity: Activity[];
  powerRanking: PowerRanking[];
}

export interface Competition {
  id: string;
  name: string;
  type: string;
  status: string;
  entrantCount: number;
  currentRound: string;
}

export interface CupEntrant {
  id: string;
  manager: string;
  teamName: string;
  seed: number;
  initials: string;
}

export interface CompetitionOverview {
  competitions: Competition[];
  availableEntrants: CupEntrant[];
}

/**
 * @nullable
 */
export type CupTieHome = {
  id?: string;
  manager?: string;
  teamName?: string;
  seed?: number;
  initials?: string;
} | null;

/**
 * @nullable
 */
export type CupTieAway = {
  id?: string;
  manager?: string;
  teamName?: string;
  seed?: number;
  initials?: string;
} | null;

export interface CupTie {
  id: string;
  /** @nullable */
  home: CupTieHome;
  /** @nullable */
  away: CupTieAway;
  /** @nullable */
  homeScore: number | null;
  /** @nullable */
  awayScore: number | null;
  status: string;
  /** @nullable */
  winnerId: string | null;
  /** @nullable */
  gameweek: number | null;
}

export interface CupRound {
  id: string;
  label: string;
  ties: CupTie[];
}

export type CupCompetition = Competition & {
  bracketSize: number;
  currentGameweek: number;
  rounds: CupRound[];
};

export type CompetitionInput = {
  /** @minLength 1 */
  name: string;
  type: 'knockout';
  /**
     * @minItems 4
     * @maxItems 4
     */
  entrantIds: string[];
  bracketSize: 4;
} | {
  /** @minLength 1 */
  name: string;
  type: 'knockout';
  /**
     * @minItems 8
     * @maxItems 8
     */
  entrantIds: string[];
  bracketSize: 8;
};

export interface CupAdvanceInput {
  /** @minLength 1 */
  tieId: string;
  simulate?: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface HealthStatus {
  status: string;
}