# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: leagues
    description: League and competition operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /leagues/{leagueId}/dashboard:
    get:
      operationId: getLeagueDashboard
      tags: [leagues]
      summary: Get the league dashboard
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: Dashboard snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueDashboard"
  /leagues/{leagueId}/standings:
    get:
      operationId: getLeagueStandings
      tags: [leagues]
      summary: Get traditional FPL standings
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: League standings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Standing"
  /leagues/{leagueId}/head-to-head:
    get:
      operationId: getHeadToHeadStandings
      tags: [leagues]
      summary: Get head-to-head standings
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: Head-to-head table
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HeadToHeadStanding"
  /leagues/{leagueId}/schedule:
    get:
      operationId: getHeadToHeadSchedule
      tags: [leagues]
      summary: Get the current gameweek schedule
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: Matchups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Matchup"
  /leagues:
    post:
      operationId: createLeague
      tags: [leagues]
      summary: Create a league
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeagueInput"
      responses:
        "201":
          description: Created league
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/League"
  /leagues/connect:
    post:
      operationId: connectFplLeague
      tags: [leagues]
      summary: Connect an existing FPL mini-league
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FplConnectInput"
      responses:
        "200":
          description: Connected league
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/League"
  /leagues/{leagueId}/competitions:
    get:
      operationId: getCompetitionOverview
      tags: [leagues]
      summary: Get league competitions and available cup entrants
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: League competition overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionOverview"
        "400":
          description: Invalid league
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createCompetition
      tags: [leagues]
      summary: Create a competition
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetitionInput"
      responses:
        "201":
          description: Created competition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CupCompetition"
        "400":
          description: Invalid cup setup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leagues/{leagueId}/competitions/{competitionId}:
    get:
      operationId: getCompetition
      tags: [leagues]
      summary: Get a competition bracket
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/CompetitionId"
      responses:
        "200":
          description: Competition bracket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CupCompetition"
        "400":
          description: Invalid competition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Competition not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leagues/{leagueId}/competitions/{competitionId}/advance:
    post:
      operationId: advanceCompetition
      tags: [leagues]
      summary: Resolve the next seeded cup tie
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/CompetitionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CupAdvanceInput"
      responses:
        "200":
          description: Updated competition bracket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CupCompetition"
        "400":
          description: Tie cannot be resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Competition not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    LeagueId:
      name: leagueId
      in: path
      required: true
      schema:
        type: string
    CompetitionId:
      name: competitionId
      in: path
      required: true
      schema:
        type: string
  schemas:
    League:
      type: object
      required: [id, name, season, managerCount, currentGameweek, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        season: { type: string }
        managerCount: { type: number }
        currentGameweek: { type: number }
        updatedAt: { type: string }
        fplLeagueId: { type: ["string", "null"] }
    LeagueInput:
      type: object
      required: [name, season]
      properties:
        name: { type: string, minLength: 1 }
        season: { type: string, minLength: 1 }
    FplConnectInput:
      type: object
      required: [fplLeagueId, name, season]
      properties:
        fplLeagueId: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        season: { type: string, minLength: 1 }
    Standing:
      type: object
      required: [rank, manager, teamName, gameweekPoints, totalPoints, previousRank, movement]
      properties:
        rank: { type: number }
        manager: { type: string }
        teamName: { type: string }
        gameweekPoints: { type: number }
        totalPoints: { type: number }
        previousRank: { type: number }
        movement: { type: number }
        initials: { type: string }
    HeadToHeadStanding:
      type: object
      required: [rank, manager, wins, losses, draws, pointsFor, pointsAgainst, pointDifference, leaguePoints, streak]
      properties:
        rank: { type: number }
        manager: { type: string }
        wins: { type: number }
        losses: { type: number }
        draws: { type: number }
        pointsFor: { type: number }
        pointsAgainst: { type: number }
        pointDifference: { type: number }
        leaguePoints: { type: number }
        streak: { type: string }
        initials: { type: string }
    Matchup:
      type: object
      required: [id, home, away, homeScore, awayScore, status]
      properties:
        id: { type: string }
        home: { type: string }
        away: { type: string }
        homeScore: { type: number }
        awayScore: { type: number }
        status: { type: string }
        featured: { type: boolean }
    LeagueDashboard:
      type: object
      required: [league, topScorer, activity, powerRanking]
      properties:
        league: { $ref: "#/components/schemas/League" }
        topScorer:
          $ref: "#/components/schemas/Standing"
        activity:
          type: array
          items: { $ref: "#/components/schemas/Activity" }
        powerRanking:
          type: array
          items: { $ref: "#/components/schemas/PowerRanking" }
    Activity:
      type: object
      required: [id, title, detail, time, kind]
      properties:
        id: { type: string }
        title: { type: string }
        detail: { type: string }
        time: { type: string }
        kind: { type: string }
    PowerRanking:
      type: object
      required: [rank, manager, movement, score]
      properties:
        rank: { type: number }
        manager: { type: string }
        movement: { type: number }
        score: { type: number }
    Competition:
      type: object
      required: [id, name, type, status, entrantCount, currentRound]
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string }
        status: { type: string }
        entrantCount: { type: number }
        currentRound: { type: string }
    CompetitionOverview:
      type: object
      required: [competitions, availableEntrants]
      properties:
        competitions:
          type: array
          items: { $ref: "#/components/schemas/Competition" }
        availableEntrants:
          type: array
          items: { $ref: "#/components/schemas/CupEntrant" }
    CupEntrant:
      type: object
      required: [id, manager, teamName, seed, initials]
      properties:
        id: { type: string }
        manager: { type: string }
        teamName: { type: string }
        seed: { type: number }
        initials: { type: string }
    CupTie:
      type: object
      required: [id, home, away, homeScore, awayScore, status, winnerId, gameweek]
      properties:
        id: { type: string }
        home:
          type: ["object", "null"]
          additionalProperties: false
          properties:
            id: { type: string }
            manager: { type: string }
            teamName: { type: string }
            seed: { type: number }
            initials: { type: string }
        away:
          type: ["object", "null"]
          additionalProperties: false
          properties:
            id: { type: string }
            manager: { type: string }
            teamName: { type: string }
            seed: { type: number }
            initials: { type: string }
        homeScore: { type: ["number", "null"] }
        awayScore: { type: ["number", "null"] }
        status: { type: string }
        winnerId: { type: ["string", "null"] }
        gameweek: { type: ["number", "null"] }
    CupRound:
      type: object
      required: [id, label, ties]
      properties:
        id: { type: string }
        label: { type: string }
        ties:
          type: array
          items: { $ref: "#/components/schemas/CupTie" }
    CupCompetition:
      allOf:
        - $ref: "#/components/schemas/Competition"
        - type: object
          required: [bracketSize, currentGameweek, rounds]
          properties:
            bracketSize: { type: number }
            currentGameweek: { type: number }
            rounds:
              type: array
              items: { $ref: "#/components/schemas/CupRound" }
    CompetitionInput:
      oneOf:
        - type: object
          required: [name, type, entrantIds, bracketSize]
          properties:
            name: { type: string, minLength: 1 }
            type: { type: string, enum: [knockout] }
            entrantIds:
              type: array
              uniqueItems: true
              minItems: 4
              maxItems: 4
              items: { type: string }
            bracketSize: { type: number, const: 4 }
        - type: object
          required: [name, type, entrantIds, bracketSize]
          properties:
            name: { type: string, minLength: 1 }
            type: { type: string, enum: [knockout] }
            entrantIds:
              type: array
              uniqueItems: true
              minItems: 8
              maxItems: 8
              items: { type: string }
            bracketSize: { type: number, const: 8 }
    CupAdvanceInput:
      type: object
      required: [tieId]
      properties:
        tieId: { type: string, minLength: 1 }
        simulate: { type: boolean }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    HealthStatus:
      type: object
      properties:
        status: { type: string }
      required: [status]