# lib/api-zod/src/generated/types/competitionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type CompetitionInput = {
  /** @minLength 1 */
  name: string;
  type: 'knockout';
  /**
     * @minItems 4
     * @maxItems 4
     */
  entrantIds: string[];
  bracketSize: 4;
} | {
  /** @minLength 1 */
  name: string;
  type: 'knockout';
  /**
     * @minItems 8
     * @maxItems 8
     */
  entrantIds: string[];
  bracketSize: 8;
};