# lib/api-zod/src/generated/types/competitionOverview.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Competition } from './competition';
import type { CupEntrant } from './cupEntrant';

export interface CompetitionOverview {
  competitions: Competition[];
  availableEntrants: CupEntrant[];
}