# lib/api-zod/src/generated/types/cupAdvanceInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CupAdvanceInput {
  /** @minLength 1 */
  tieId: string;
  simulate?: boolean;
}