# lib/api-zod/src/generated/types/cupCompetition.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Competition } from './competition';
import type { CupRound } from './cupRound';

export type CupCompetition = Competition & {
  bracketSize: number;
  currentGameweek: number;
  rounds: CupRound[];
};