# lib/api-zod/src/generated/types/cupEntrant.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CupEntrant {
  id: string;
  manager: string;
  teamName: string;
  seed: number;
  initials: string;
}