# lib/api-zod/src/generated/types/cupRound.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CupTie } from './cupTie';

export interface CupRound {
  id: string;
  label: string;
  ties: CupTie[];
}