# lib/api-zod/src/generated/types/cupTie.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CupTieAway } from './cupTieAway';
import type { CupTieHome } from './cupTieHome';

export interface CupTie {
  id: string;
  /** @nullable */
  home: CupTieHome;
  /** @nullable */
  away: CupTieAway;
  /** @nullable */
  homeScore: number | null;
  /** @nullable */
  awayScore: number | null;
  status: string;
  /** @nullable */
  winnerId: string | null;
  /** @nullable */
  gameweek: number | null;
}